Try to remove SET_COMPILE_OUTPUTS_MODIFIERS macro
The macro was used to change modifiers on compile outputs (namely object files).
The only use of macro was to add `norel` modifier, which changes the way how output paths are constructed.
- W/o the modifier (default case) output path is always localized to module and disambiguated by use of `__` for `..` and `_` for `.` in path to file relative to module (the latter rule is only applied when file is in subdirectory to disambiguate outputs from same input in nested modules).
- W/ modifier output path is exact match to input path (only changing `$S` to `$B` if input is from source tree).

Generally norel is dangerous but needed in 2 major cases
- The output is produced by tool which lacks ability to control output path and always deduce output path from input (protoc being prominent example)
- The input is not directly in module and output is used in source code (not consumed by build system) e.g. header file generation from input in subdirectory.

The macro was introduced along with strict disambiguation for outputs to cure problematic cases. However it is now used only in one case and seemingly to just contract path on Windows (disambiguation resulted in lengthy path due to distant relative references in aforementioned module).  Build system now supports long paths on Windows, so this case shouldn't be a problem any more.

ref:9478175
sync: https://proxy.sandbox.yandex-team.ru/3115643340